Update repositories to use single data source
