+++ app/src/main/java/com/xinto/mauth/Mauth.kt
@@ -21,7 +21,7 @@ class Mauth : Application() {
                 qrScanModule,
                 dbModule,
                 otpModule,
-                datastoreModule,
+                preferencesModule,
             )
         }
     }
