+++ app/src/main/java/com/xinto/mauth/domain/repository/MainRepository.kt
@@ -1,10 +1,7 @@
 package com.xinto.mauth.domain.repository
 
-import androidx.datastore.core.DataStore
-import androidx.datastore.preferences.core.Preferences
-import androidx.datastore.preferences.core.booleanPreferencesKey
+import com.xinto.mauth.preferences.PreferenceManager
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.map
 
 interface MainRepository {
 
@@ -13,17 +10,11 @@ interface MainRepository {
 }
 
 class MainRepositoryImpl(
-    private val dataStore: DataStore<Preferences>
+    private val preferenceManager: PreferenceManager
 ) : MainRepository {
 
     override fun observeSecureMode(): Flow<Boolean> {
-        return dataStore.data.map {
-            it[KEY_PRIVATE_MODE] ?: false
-        }
-    }
-
-    private companion object {
-        val KEY_PRIVATE_MODE = booleanPreferencesKey("private_mode")
+        return preferenceManager.observePrivateMode()
     }
 
 }
