+++ app/src/main/java/com/xinto/mauth/domain/repository/SettingsRepository.kt
@@ -1,11 +1,7 @@
 package com.xinto.mauth.domain.repository
 
-import androidx.datastore.core.DataStore
-import androidx.datastore.preferences.core.Preferences
-import androidx.datastore.preferences.core.booleanPreferencesKey
-import androidx.datastore.preferences.core.edit
+import com.xinto.mauth.preferences.PreferenceManager
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.map
 
 interface SettingsRepository {
 
@@ -15,23 +11,15 @@ interface SettingsRepository {
 }
 
 class SettingsRepositoryImpl(
-    private val dataStore: DataStore<Preferences>
+    private val preferenceManager: PreferenceManager
 ) : SettingsRepository {
 
     override fun observePrivateMode(): Flow<Boolean> {
-        return dataStore.data.map {
-            it[KEY_PRIVATE_MODE] ?: false
-        }
+        return preferenceManager.observePrivateMode()
     }
 
     override suspend fun setPrivateMode(value: Boolean) {
-        dataStore.edit {
-            it[KEY_PRIVATE_MODE] = value
-        }
-    }
-
-    private companion object {
-        val KEY_PRIVATE_MODE = booleanPreferencesKey("private_mode")
+        preferenceManager.setPrivateMode(value)
     }
 
 }
